+++ usr.sbin/bgpd/session.h
@@ -1,4 +1,4 @@
-/*	$OpenBSD: session.h,v 1.196 2026/05/07 09:17:27 claudio Exp $ */
+/*	$OpenBSD: session.h,v 1.197 2026/05/12 20:27:31 claudio Exp $ */
 
 /*
  * Copyright (c) 2003, 2004 Henning Brauer <henning@openbsd.org>
@@ -340,7 +340,7 @@ struct peer	*getpeerbyid(struct bgpd_config *, uint32_t);
 void		 session_handle_update(struct peer *, struct ibuf *);
 void		 session_handle_rrefresh(struct peer *, struct route_refresh *);
 void		 session_graceful_restart(struct peer *);
-void		 session_graceful_flush(struct peer *, uint8_t, const char *);
+void		 session_graceful_flush(struct peer *, u_int, const char *);
 void		 session_mrt_dump_state(struct peer *);
 void		 session_mrt_dump_bgp_msg(struct peer *, struct ibuf *,
 		    enum msg_type, enum direction);
